frontend/src-tauri/python/requirements-diarization.txt
# Speaker Diarization Dependencies
# Install with: pip install -r requirements-diarization.txt

# Core dependencies
pyannote.audio>=3.1.0
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0

# Additional dependencies for pyannote
speechbrain>=0.5.0
asteroid-filterbanks>=0.4.0
torch-audiomentations>=0.11.0
pytorch-lightning>=2.0.0

# Hugging Face integration
huggingface-hub>=0.16.0
transformers>=4.30.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Optional: ONNX Runtime for optimized inference
# onnxruntime-gpu>=1.15.0  # For GPU
# onnxruntime>=1.15.0      # For CPU
